Fix  language detection for Widows
