+++ shared/locales/de.json
@@ -1,3 +1,3 @@
 {
-    "cli.command.signup": "{cmd} | Create a new YeetFile account"
+    "cli.command.signup": "{cmd} | Ein neues Yeetfile Konto anlegen"
 }
